Make sure private key share is tECDSA compatible during marshaling
